examples/tutorial/requirements.txt
aiosqlite==0.19.0         # Async version of sqlite
cannula[codegen]==0.20.0  # The reason you are reading this
click==8.1.7              # Command line tool helper
fastapi[all]==0.115.0     # The main asgi application logic
greenlet==3.0.3           # For sqlalchemy async support
ruff==0.1.4               # For formatting the code
sqlalchemy==2.0.25        # For database table access
uvicorn==0.25.0           # Web server for asgi applications
python-dateutil==2.9.0    # Parse date time objects

# Testing Tools
pytest<8
pytest-asyncio
pytest-cov
pytest-httpx
pytest-mock
mypy
